Remove superfluous assignment to '$arg{perm}'.
In this instance of an invocation of 'lstat' -- unlike the other instances
later on within _rmtree() -- we are only consulting the first two elements in
the list returned by that builtin function.  Hence, we are only assigning to
'device' and 'inode'; 'perm' will always be undefined.  Within _rmtree()
'$arg{perm}' is never consulted.  Hence, the assignment is superfluous and
should be removed.