Add tests for previously uncovered source code.
t/Path.t:

Test case of 0 arguments for remove_tree() and rmtree().
Demonstrate that carp message is emitted.

Test cases where first arg to rmtree() is undefined value or empty string.
Show that we get the expected _carp message and return value.

Test cases of 0 arguments for rmtree(), make_path() and mkpath().

mkpath():  For each of the 'user' and 'uid' arguments, demonstrate that we log
an error message as expected.  However, this exposes a problem.
Notwithstanding a phony value for either 'user' or 'uid', mkpath() plows ahead
and creates the directories requested.  This behavior is probably
long-standing -- but is it desirable.

lib/File/Path.pm: Consolidate variable definitions to keep Devel::Cover happy.