Test behavior on Win32 for point at which getpwnam and getgrnam were first used.
This branch was created at the commit at which getpwnam() and getgrnam were
first used within the definition of mkpath().  We have inferred that the
places where those built-ins are used are not meaningful on Win32 but those
cases were not explicitly tested.

The tests in this file provide either 'owner' or 'group' elements in the
hashref passed as the final argument to make_path() -- which internally calls
mkpath().  Doing so should trigger branches within the code that call
getpwnam() and getgranam(), respectively, resulting in an error condition when
run on Win32, as those built-ins are not implemented on that OS.  We want to
see what is found in $@ at these places.  We also want to see if we trap
anything via error => \$error.