Make subdirectory names used in tests more random.
Normally when creating a temporary directory for testing, we would use
File::Temp::tempdir() to get a path like this:

    $ perl -MFile::Temp=tempdir -E '$t=tempdir(); say $t;'
    /tmp/PF2MzK1R_t

But File::Temp calls File::Path::rmtree() internally.  So we resorted to
File::Spec::Functions::tmpdir(), which in most cases will simply return:

    $ perl -MFile::Spec::Functions=tmpdir -E '$t=tmpdir(); say $t;'
    /tmp

But we overlooked the fact that a user might already have directories named
like this:

    /tmp/d
    /tmp/e
    /tmp/f

This caused a test failure on one of Nigel Horne's CPANtesters rigs:
http://www.cpantesters.org/cpan/report/b2a64ee8-4703-11e7-be6b-64b9d07fe62d

We're addressing this by switching to 12-letter strings generated by
String::PerlIdentifier and applied manually.