Change method of generating names for non-existent user and group.
We issue a warning when a directory is created owned by a non-existent user or
group.  To test that that feature works, we have to create names for a phony
user and a phony group.

Until now, the method used to use getpwnam() to search for the name of the
user with the highest UID and to use getgrnam() to search for the name of the
group with the highest GID.  Then, that name would be autoincremented like
this:

        ++$max_user;

... where the intent is, e.g., to take the string 'nobody' and change it to
'nobodz'.  But if, say, the user with the highest UID is 'system-dbus-proxy',
that won't work because it runs afoul of this provision described in 'perldoc
perlop' ("Auto-increment and Auto-decrement"):

    "The auto-increment operator has a little extra builtin magic to it. If you
    increment a variable that is numeric, or that has ever been used in a
    numeric context, you get a normal increment. If, however, the variable has
    been used in only string contexts since it was set, and has a value that
    is not the empty string and matches the pattern "/^[a-zA-Z]*[0-9]*\z/",
    the increment is done as a string, preserving each character within its
    range, with carry[.]"

A string like 'system-dbus-proxy' -- which in certain environments is a
plausible user name -- does not match that pattern and the autoincrement
statement ends up returning 1, which causes tests in t/Path_root.t to fail.

So we have to devise a new way of generating names for a phony user and phony
group.

For: https://rt.cpan.org/Ticket/Display.html?id=121967

Thanks to zdm and NICKY for reporting this problem.